ModuleSceneBase Added + General Cleaning

+++ OOP_Version/Source/ComponentCamera.cpp
@@ -7,7 +7,7 @@
 
 ComponentCamera::ComponentCamera(Entity* entity) : position(0.0f), rotation(0.0f), Component(Component::Type::CAMERA, entity)
 {
-	UpdateProjectionMatrix(0, App->window->GetWidth(), App->window->GetHeight(), 0);
+	UpdateProjectionMatrix(0, (float)App->window->GetWidth(), (float)App->window->GetHeight(), 0);
 	UpdateViewMatrix();
 }
